+++ packages/config/src/Devices.ts
@@ -4,15 +4,17 @@ import {
 	JSONObject,
 	ObjectKeyMap,
 	ReadonlyObjectKeyMap,
+	stringify,
 } from "@zwave-js/shared";
 import { entries } from "alcalzone-shared/objects";
 import { isArray, isObject } from "alcalzone-shared/typeguards";
-import { pathExists, readFile } from "fs-extra";
+import { pathExists, readFile, writeFile } from "fs-extra";
 import JSON5 from "json5";
 import path from "path";
 import log from "./Logger";
 import {
 	configDir,
+	formatId,
 	getDeviceEntryPredicate,
 	hexKeyRegex4Digits,
 	throwInvalidConfig,
@@ -37,7 +39,7 @@ export type ParamInfoMap = ReadonlyObjectKeyMap<
 >;
 
 const indexPath = path.join(configDir, "devices/index.json");
-let index: readonly DeviceConfigIndexEntry[] | undefined;
+let index: DeviceConfigIndexEntry[] | undefined;
 
 /** @internal */
 export async function loadDeviceIndexInternal(): Promise<typeof index> {
@@ -82,6 +84,60 @@ export async function loadDeviceIndex(): Promise<void> {
 			throw e;
 		}
 	}
+}
+
+export async function writeIndexToFile(): Promise<void> {
+	if (!index) {
+		throw new ZWaveError(
+			"The config has not been loaded yet!",
+			ZWaveErrorCodes.Driver_NotReady,
+		);
+	}
+
+	await writeFile(indexPath, stringify(index));
+}
+
+/**
+ * Looks up the definition of a given device in the configuration DB
+ * @param manufacturerId The manufacturer id of the device
+ * @param productType The product type of the device
+ * @param productId The product id of the device
+ * @param filename The path to the json configuration of this device
+ * @param firmwareVersionMin Min firmware version
+ * @param firmwareVersionMax Max firmware version
+ *
+ */
+export function addDeviceToIndex(
+	manufacturerId: number,
+	productType: number,
+	productId: number,
+	filename: string,
+	firmwareVersionMin?: string,
+	firmwareVersionMax?: string,
+): void {
+	if (!index) {
+		throw new ZWaveError(
+			"The config has not been loaded yet!",
+			ZWaveErrorCodes.Driver_NotReady,
+		);
+	}
+	// Look up the device in the index
+	const indexEntry: DeviceConfigIndexEntry = {
+		manufacturerId: formatId(manufacturerId),
+		productId: formatId(productId),
+		productType: formatId(productType),
+		firmwareVersion: {
+			min: firmwareVersionMin || "0.0",
+			max: firmwareVersionMax || "255.255",
+		},
+		filename: filename,
+	};
+
+	index.push(indexEntry);
+}
+
+export function getIndex(): DeviceConfigIndexEntry[] | undefined {
+	return index;
 }
 
 /**
@@ -439,6 +495,18 @@ error in compat option queryOnWakeup`,
 					}),
 			) as any;
 		}
+
+		if (definition.keepS0NonceUntilNext != undefined) {
+			if (definition.keepS0NonceUntilNext !== true) {
+				throwInvalidConfig(
+					"devices",
+					`config/devices/${filename}:
+error in compat option keepS0NonceUntilNext`,
+				);
+			}
+
+			this.keepS0NonceUntilNext = definition.keepS0NonceUntilNext;
+		}
 	}
 
 	public readonly queryOnWakeup?: [
@@ -451,6 +519,8 @@ error in compat option queryOnWakeup`,
 			| Pick<ValueID, "property" | "propertyKey">
 		)[]
 	][];
+
+	public keepS0NonceUntilNext?: boolean;
 }
 
 export class ParamInformation {