+++ packages/zwave-js/src/lib/node/Node.ts
@@ -1697,6 +1697,10 @@ version:               ${this.version}`;
 			ZWaveErrorCodes.Controller_MessageDropped,
 		);
 
+		// And also delete all previous nonces we sent the node since they
+		// should no longer be used
+		this.driver.securityManager.deleteAllNoncesForReceiver(this.id);
+
 		// Now send the current nonce
 		try {
 			await this.commandClasses.Security.sendNonce();
@@ -1718,11 +1722,14 @@ version:               ${this.version}`;
 
 		secMan.setNonce(
 			{
-				nodeId: this.id,
+				issuer: this.id,
 				nonceId: secMan.getNonceId(command.nonce),
 			},
-			command.nonce,
-			true,
+			{
+				nonce: command.nonce,
+				receiver: this.driver.controller.ownNodeId!,
+			},
+			{ free: true },
 		);
 	}
 
@@ -2515,7 +2522,7 @@ version:               ${this.version}`;
 
 		// Refresh the get timeout
 		if (this._firmwareUpdateStatus.getTimeout) {
-			console.warn("refreshed get timeout");
+			// console.warn("refreshed get timeout");
 			this._firmwareUpdateStatus.getTimeout = this._firmwareUpdateStatus.getTimeout
 				.refresh()
 				.unref();